opnode: Do not drop all deposits on parsing error (#383)
* Change deposit gasLimit from uint256 to uint64

This prevents the user from setting a gas limit that the rollup node
is not able to parse.

* opnode: Do not drop all deposits on parsing error

Only drop the affected deposit. This does mean loss of funds, but
this is also a case that should never happen. In addition, halting
the chain means that while funds are safe, we are exposed to a
denial of service attack if it would be possible to cause a parsing
error from the deposit feed contract.

+++ packages/contracts/contracts/L1/abstracts/DepositFeed.sol
@@ -33,7 +33,7 @@ abstract contract DepositFeed {
         address indexed to,
         uint256 mint,
         uint256 value,
-        uint256 gasLimit,
+        uint64 gasLimit,
         bool isCreation,
         bytes data
     );
@@ -54,7 +54,7 @@ abstract contract DepositFeed {
     function depositTransaction(
         address _to,
         uint256 _value,
-        uint256 _gasLimit,
+        uint64 _gasLimit,
         bool _isCreation,
         bytes memory _data
     ) public payable {
