+++ packages/integration-tests/contracts/MultiDepositor.sol
@@ -4,7 +4,7 @@ interface DepositFeed {
     function depositTransaction(
         address _to,
         uint256 _value,
-        uint256 _gasLimit,
+        uint64 _gasLimit,
         bool _isCreation,
         bytes memory _data
     ) external payable;
